app/requirements.txt
# Core
streamlit>=1.32.0

# DICOM
pydicom>=2.4.0

# Image processing
numpy>=1.26.0
scipy>=1.12.0
Pillow>=10.0.0
matplotlib>=3.8.0

# LLM APIs
openai>=1.20.0
google-generativeai>=0.5.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
